Initial commit of the script to connect LeadDocket with BigQuery
In this commit, the script is only half-finished. More work needs
to be done to integrate with Google Secrets Manager, upsert needs
to be implemented for the staging table, and the code needs to be
optimized.

Additional cleanup is necessary to remove comments, delete notes,
and document the existing functions.