Fixed a schema error, fixed a typo in the README
This commit fixes a bug introduced in the refactor. The name for some
fields were not being accessed with proper capitalization and therefore
the script failed to load those fields.

This commit also fixes a typo in the README where the prod_table_id was
listed twice instead of the staging_table_id.